[Code] Bit more fine-grained configuration for invalid typespec
Now `*signal-on-invalid-typespes*` affects only cases that the typespec
is invalid according to ANSI, but (possibly) can be handled and default
`lisp:typep` doing so. That is, a) compound-only typespecs being a symbol,
or b) `satisfies` with lambda-expression.
Other invalid typespecs like not a symbol nor a list will be error of
type `bad-type-specifier` immediately.