[Code] Add utility macro `named-lambda`
Returns a function as `lambda` but it has NAME.
There's no use of NAME other than it is printed as "#<lexical-closure: NAME>"
to help identify what function that is.

NAME can be anything, not restricted to symbol.
Function created by `named-lambda` doesn't have implicit block named NAME,
so do it on your own if necessary. (Remember that `block` tag must be a
symbol :p)