Add accessor `ansify.ext:function-info` #code
Provides way to store data on function object in same way as hash-table.

    (setf (ansify.ext:function-info #'SOME-FN :KEY) WHATEVER)

It stores the data on function object itself, not on symbol naming the
function. So you never miss the data.
To achieve this, it modifies lambda-list part of `si:closure-body` and
put a hash-table as INIT-FORM of &aux parameter.