Modify structure printer for `*method`: specializer in form you write
`*method` holds specializers in different form (for ease of check) than
what you write in `defmethod` form, and this internal form was printed
before. As a result, the output is a) confusing, and b) bit too long.

So now it prints in a form both compact and we are familiar with.