`ensure-generic-function` updates the GF
Only `:documentation` and `:lambda-list` is supported. (other options
are not supported on GF in the first place.)

The GF thunk (made by `%make-generic-function`) is modified to get
narg-min/max from GF-INFO on invocation instead of having them in
lexenv. (That means it parse the lambda list on every invocation.
Might be slow.)